refactor(person): 파일 이동, final 정의
