+++ src/main/java/blackjack/domain/player/Person.java
@@ -1,10 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 public abstract class Person {
 
     private final String name;
 
-    public Person(String name) {
+    public Person(final String name) {
         this.name = name;
     }
