+++ src/test/java/blackjack/domain/DealerTest.java
@@ -1,8 +1,11 @@
 package blackjack.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
+import blackjack.domain.card.CardDeck;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.state.State;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,10 +14,10 @@ class DealerTest {
     @DisplayName("딜러_생성_테스트")
     @Test
     void createDealerTest() {
-        Dealer dealer = new Dealer("딜러", new State(CardDeck.pop(2)));
+        Dealer dealer = new Dealer("딜러", new State(CardDeck.pop(2), true));
         assertAll(
-            ()->assertThat(dealer.getName()).isEqualTo("딜러"),
-            ()->assertThat(dealer).isNotNull()
+            () -> assertThat(dealer.getName()).isEqualTo("딜러"),
+            () -> assertThat(dealer).isNotNull()
         );
     }
 }
